# swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    { "url": "https://kapusta-srv.herokuapp.com", "description": "", "variables": {} },
    { "url": "http://localhost:8080", "description": "", "variables": {} }
  ],
  "info": {
    "version": "1.0.0",
    "title": "Kapusta API",
    "description": "Financial API for you home budget planning",
    "termsOfService": ""
  },

  "tags": [
    {
      "name": "auth",
      "description": "Authorization"
    },
    {
      "name": "transactions",
      "description": "financial operations"
    }
  ],
  "paths": {
    "/user/incomes": {
      "get": {
        "tags": ["transactions"],
        "summary": "get incomes",
        "operationId": "getincomes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "get incomes",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Incomes" },
                "example": {
                  "user": {
                    "operations": {
                      "incomes": [{ "date": "Март", "amount": 15 }]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": ["transactions"],
        "summary": "patch incomes",
        "operationId": "patchincomes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "patch incomes",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/200" },
                "example": { "categoryTotal": 9 }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "date": "03-15-2021",
                  "category": "ЗП",
                  "description": "зарплата",
                  "amount": "9",
                  "id": "1"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["transactions"],
        "summary": "delete incomes",
        "operationId": "deleteincomes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "delete incomes",
            "content": { "text/html": { "schema": {} } }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "id": "1"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/costs": {
      "get": {
        "tags": ["transactions"],
        "summary": "get costs",
        "operationId": "getcosts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "get costs",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Costs" },
                "example": {
                  "user": {
                    "operations": {
                      "costs": [{ "date": "Март", "amount": 30 }]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": ["transactions"],
        "summary": "patch costs",
        "operationId": "patchcosts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "patch costs",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/200" },
                "example": { "categoryTotal": 30 }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "date": "03-16-2021",
                  "category": "Продукты",
                  "description": "Молоко",
                  "amount": "20",
                  "id": "2"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["transactions"],
        "summary": "delete costs",
        "operationId": "deletecosts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "delete costs",
            "content": { "text/html": { "schema": {} } }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "id": "1"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/information/{date}": {
      "get": {
        "tags": ["transactions"],
        "summary": "get information",
        "operationId": "getinformation",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "date",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{2}-\\d{2}-\\d{4}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Information" },
                "example": {
                  "costs": {
                    "total": 30,
                    "Продукты": { "Макароны": 10, "Молоко": 20, "total": 30 }
                  },
                  "incomes": {
                    "total": 15,
                    "ЗП": { "зарплата": 15, "total": 15 }
                  },
                  "costsArr": [
                    { "total": 30, "name": "алкоголь" },
                    { "total": 150, "name": "продукты" }
                  ],
                  "incomesArr": [{ "total": 12000, "name": "зарплата" }]
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "register user",
        "operationId": "registeruser",
        "parameters": [],
        "responses": {
          "200": { "description": "", "headers": {} },
          "201": {
            "description": "register user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Login" },
                "example": {
                  "id": "604fac1d57822d2444d93cc2",
                  "email": "test22@gmail.com",
                  "name": "test"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "email": "test22@gmail.com",
                  "password": "1231231212",
                  "name": "test"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "login user",
        "operationId": "loginuser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "login user",
            "content": {
              "application/json": {
                "example": {
                  "token": "Access token",
                  "refreshToken": "Refresh token",
                  "user": {
                    "email": "shabliioleg@gmail.com",
                    "name": "Xoms",
                    "avatarURL": "https://res.cloudinary.com/dfmmznnzy/image/upload/v1616238499/peife2ztl9lhtqtdpsdi.png",
                    "balance": 0
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "email": "shabliioleg@gmail.com",
                  "password": "password"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "logout user",
        "operationId": "logoutuser",
        "parameters": [],
        "responses": { "200": { "description": "", "headers": {} } },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOiI2MDU1ZDdhNDQyZTM3ZTIyYmNhYjIzNDYiLCJpYXQiOjE2MTYyODM2ODV9.Oc1LjpsYXCb4D_I7Fn4l5svJs8LwotdytKqXSrvlcT8"
                }
              }
            }
          }
        }
      }
    },
    "auth/mail-verify/:token": {
      "get": {
        "tags": ["auth"],
        "summary": "Email verification endpoint, accept Verification Token as a parameter",
        "parameters": [{
          "name": "token",
          "in": "path",
          "description": "Email verification token",
          "required": true
           }],
        "responses": { "200": { "description": "Return html markup with link to FE and set isVerified to 'true' in current User entity"} }
      }
    },
    "auth/refresh-token": {
      "post": {
        "tags": ["auth"],
        "summary": "Returns new accessToken",
        "responses": {
          "200": {
            "description": "New access token",
            "content": {
              "application/json": {
                "example": {
                  "token": "Access token"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOiI2MDU1ZDdhNDQyZTM3ZTIyYmNhYjIzNDYiLCJpYXQiOjE2MTYyODM2ODV9.Oc1LjpsYXCb4D_I7Fn4l5svJs8LwotdytKqXSrvlcT8"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Incomes": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "operations": {
                "type": "object",
                "properties": {
                  "incomes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "date": {
                          "type": "string",
                          "description": "",
                          "default": "Март"
                        },
                        "amount": {
                          "type": "integer",
                          "description": "",
                          "default": 15
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Costs": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "operations": {
                "type": "object",
                "properties": {
                  "costs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "date": {
                          "type": "string",
                          "description": "",
                          "default": "Март"
                        },
                        "amount": {
                          "type": "integer",
                          "description": "",
                          "default": 15
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Information": {
        "type": "object",
        "properties": {
          "costs": {
            "type": "object",
            "properties": {
              "total": {
                "type": "string",
                "description": "",
                "default": ""
              },
              "category": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "string",
                      "description": "",
                      "default": ""
                    },
                    "description": {
                      "type": "string",
                      "description": "",
                      "default": ""
                    }
                  }
                }
              }
            }
          },
          "incomes": {
            "type": "object",
            "properties": {
              "total": {
                "type": "string",
                "description": "",
                "default": ""
              },
              "category": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "string",
                      "description": "",
                      "default": ""
                    },
                    "description": {
                      "type": "string",
                      "description": "",
                      "default": ""
                    }
                  }
                }
              }
            }
          },
          "incomesArr": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "string",
                  "description": "",
                  "default": ""
                },
                "name": {
                  "type": "string",
                  "description": "",
                  "default": ""
                }
              }
            }
          },
          "costsArr": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "string",
                  "description": "",
                  "default": ""
                },
                "name": {
                  "type": "string",
                  "description": "",
                  "default": ""
                }
              }
            }
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "",
            "default": "604fac1d57822d2444d93cc2"
          },
          "email": {
            "type": "string",
            "description": "",
            "default": "test22@gmail.com"
          },
          "name": { "type": "string", "description": "", "default": "test" }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
  "externalDocs": { "url": "", "description": "" },
  "warnings": []
}